[feat] 댓글 삭제 - 권한 설정 (확인 필요)

+++ src/main/resources/static/js/app/index.js
@@ -110,20 +110,26 @@ var main = {
     commentDelete : function (commentId) {
         var data = {
             postsId: $('#postsId').val(),
-            // commentId: $('#commentId').val()
+            postsUser: $('#postsUser').val(),
+            user: $('#commentUser').val()
+        }
+        if (data.user === data.postsUser) {
+            $.ajax({
+                type: 'DELETE',
+                url: '/api/v1/posts/'+ data.postsId + '/comments/'+commentId,
+                dataType: 'json',
+                contentType:'application/json; charset=utf-8'
+            }).done(function() {
+                alert('댓글이 삭제되었습니다.');
+                window.location.reload();
+            }).fail(function (error) {
+                alert(JSON.stringify(error));
+            });
+        } else {
+            alert(data.user+', '+data.postsUser);
+            alert('댓글 삭제 권한이 없습니다.');
         }
 
-        $.ajax({
-            type: 'DELETE',
-            url: '/api/v1/posts/'+ data.postsId + '/comments/'+commentId,
-            dataType: 'json',
-            contentType:'application/json; charset=utf-8'
-        }).done(function() {
-            alert('댓글이 삭제되었습니다.');
-            window.location.reload();
-        }).fail(function (error) {
-            alert(JSON.stringify(error));
-        });
     },
